requirements.txt
Django==3.2
environs==9.3.5
Pillow==8.4.0
django-admin-sortable2==1.0.2

